DSPy/requirements.txt
dspy-ai
boto3